+++ ktlint-rule-engine/src/main/kotlin/com/pinterest/ktlint/rule/engine/internal/KotlinPsiFileFactory.kt
@@ -103,6 +103,7 @@ private class LoggerFactory : DiagnosticLogger.Factory {
                 message: String?,
                 t: Throwable?,
             ) {}
+
             override fun error(
                 message: String?,
                 vararg details: String?,